+++ js/create-array-features.js
@@ -1,0 +1,8 @@
+import { getRandomInRange } from './rundom-number.js';
+
+const getRandomArrayFeatures = (array) => {
+  const numElements = getRandomInRange(0, array.length - 1);
+  return array.slice(0, numElements);
+};
+
+export { getRandomArrayFeatures };
