+++ js/random-array-element.js
@@ -1,0 +1,4 @@
+import { getRandomInRange } from './rundom-number.js';
+const getRandomArrayElement = (array) => array[getRandomInRange(0, array.length - 1)];
+
+export { getRandomArrayElement };
